+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 * Also support `circulation` `11.0`. Refs UIPCIR-21.
+* Receiving: Add filters for Expected receipt date and Received date. Refs UIREC-120.
 
 ## [1.3.3](https://github.com/folio-org/ui-receiving/tree/v1.3.3) (2021-04-21)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.3.2...v1.3.3)